requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
pytest
flask-marshmallow
marshmallow-sqlalchemy
gunicorn
marshmallow-jsonapi
flask-cors
pgeocode
coverage
